Перенос хуков карты в отдельный каталог
