+++ project/src/hooks/index.ts
@@ -1,0 +1,7 @@
+import { useMap } from './map/use-map';
+import { useOffers } from './map/use-offers';
+import { useLocation } from './map/use-location';
+import { useCurrentMarker } from './map/use-current-marker';
+
+
+export { useMap, useOffers, useLocation, useCurrentMarker };
