+++ project/src/hooks/map/use-location.ts
@@ -1,0 +1,13 @@
+import { useEffect } from 'react';
+import leaflet, { Map } from 'leaflet';
+
+import { Location } from '../../types/location';
+
+
+export const useLocation = (map: Map | null, location: Location) => {
+  useEffect(() => {
+    if (map) {
+      map.setView(new leaflet.LatLng(location.latitude, location.longitude), location.zoom);
+    }
+  }, [map, location]);
+};
